requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
fastapi
uvicorn
requests
edge-tts
qdrant-client
fastembed
